Move forward declarations into separate headers
Reverts the previous change from typedef -> class
for OrbitalVector etc.
