+++ src/chemistry/chemistry.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace mrchem {
+
+class Nucleus;
+class Nuclei;
+class Molecule;
+class Element;
+class PeriodicTable;
+
+} //namespace mrchem
