+++ src/initial_guess/sad.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "qmfunctions.h"
+#include "chemistry.h"
+
 /** @file sad.h
  *
  * @brief Module for generating initial guess as superposition of atomic densities
@@ -9,10 +12,6 @@
  */
 
 namespace mrchem {
-
-class OrbitalVector;
-class Molecule;
-
 namespace initial_guess {
 namespace sad {
