+++ src/mrchem.h
@@ -37,21 +37,19 @@ namespace MATHCONST {
     const double sqrt_pi =    1.7724538509055160273;
 }
 
-namespace SPIN { enum type { Paired, Alpha, Beta }; }
-namespace NUMBER { enum type { Total, Real, Imag }; }
-namespace DENSITY { enum type { Total, Spin, Alpha, Beta }; }
+using ComplexInt = std::complex<int>;
+using ComplexDouble = std::complex<double>;
 
-typedef std::complex<double> ComplexDouble;
-typedef std::function<double (const double *r)> DoubleFunction;
-typedef std::function<ComplexDouble (const double *r)> ComplexFunction;
+using DoubleFunction = std::function<double (const double *r)>;
+using ComplexFunction = std::function<ComplexDouble (const double *r)>;
 
-typedef Eigen::VectorXi  IntVector;
-typedef Eigen::VectorXd  DoubleVector;
-typedef Eigen::VectorXcd ComplexVector;
+using IntVector = Eigen::VectorXi;
+using DoubleVector = Eigen::VectorXd;
+using ComplexVector = Eigen::VectorXcd;
 
-typedef Eigen::MatrixXi  IntMatrix;
-typedef Eigen::MatrixXd  DoubleMatrix;
-typedef Eigen::MatrixXcd ComplexMatrix;
+using IntMatrix = Eigen::MatrixXi;
+using DoubleMatrix = Eigen::MatrixXd;
+using ComplexMatrix = Eigen::MatrixXcd;
 
 extern Getkw Input;
 extern mrcpp::MultiResolutionAnalysis<3> *MRA;  //< Global MRA
