+++ src/properties/properties.h
@@ -1,0 +1,15 @@
+#pragma once
+
+namespace mrchem {
+
+class SCFEnergy;
+class DipoleMoment;
+class QuadrupoleMoment;
+class Magnetizability;
+class NMRShielding;
+class HyperFineCoupling;
+class SpinSpinCoupling;
+class Polarizability;
+class OpticalRotation;
+
+} //namespace mrchem
