+++ src/qmfunctions/orbital_utils.h
@@ -1,15 +1,8 @@
 #pragma once
 
-#include "mrchem.h"
+#include "qmfunctions.h"
 
 namespace mrchem {
-
-/* The following container classes are defined as derived classes rather than
- * typedefs in order to be able to forward declare them. */
-
-class Orbital;
-class OrbitalChunk final : public std::vector<std::tuple<int, Orbital> > { };
-class OrbitalVector final : public std::vector<Orbital> { };
 
 namespace orbital {
