+++ src/qmoperators/two_electron/ExchangeOperator.h
@@ -16,8 +16,6 @@
 
 namespace mrchem {
 
-class OrbitalVector;
-
 class ExchangeOperator final : public RankZeroTensorOperator {
 public:
  ExchangeOperator(mrcpp::PoissonOperator &P, OrbitalVector &Phi, double xFac = 1.0)
