+++ src/utils/RRMaximizer.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "NonlinearMaximizer.h"
+#include "qmfunctions.h"
 
 /** subclass which defines the particular Gradient and Hessian
  * and other specific functions for a maximization of
@@ -11,8 +12,6 @@
  */
 
 namespace mrchem {
-
-class OrbitalVector;
 
 class RRMaximizer final : public NonlinearMaximizer {
 public:
